+++ cpp/utils/BaseOptions.cpp
@@ -3,6 +3,6 @@
 namespace margelo::nitro::totp {
 
 // Initialize static members of BaseSecretKeyOptions
-int BaseSecretKeyOptions::length = 12;
+int BaseSecretKeyOptions::size = 20; // 20 bytes = 32 Base32 characters (industry standard)
 
 } // namespace margelo::nitro::totp